+++ dask_geomodeling/raster/temporal.py
@@ -8,6 +8,7 @@ from pandas.tseries.frequencies import to_offset
 
 import numpy as np
 import pandas as pd
+import warnings
 
 from dask_geomodeling.utils import (
     get_dtype_max,
@@ -25,20 +26,20 @@ class Snap(RasterBlock):
     """
     Snap the time structure of a raster to that of another raster.
 
-    :param store: Return values from this block.
-    :param index: Snap values to the times from this block.
+    This operations allows to take the cell values from one raster ('store')
+    and the temporal properties of another raster ('index').
 
-    :type store: RasterBlock
-    :type index: RasterBlock
+    If the store is not a temporal raster, its cell values are copied to each
+    timestep of the index raster. If the store is also a temporal raster, this
+    operation looks at each 'index' timestamp and takes the closest 'store'
+    timestamp as cell values.
 
-    On get_data, get_meta or get_time requests, it will return the
-    data and meta from the block supplied as the store parameter, but
-    apparently have the time structure of the Store supplied as the
-    index parameter.
+    Args:
+      store (RasterBlock): Return cell values from this raster
+      index (RasterBlock): Snap values to the timestamps from this raster
 
-    In contrast to the group block, this block does not take
-    advantage of aligned bands in any way. Therefore the perfomance will
-    be noticeably worse, in particular when requesting long timeslices.
+    Returns: 
+      RasterBlock with temporal properties of the index.
     """
 
     def __init__(self, store, index):
@@ -169,17 +170,17 @@ class Snap(RasterBlock):
 
 class Shift(BaseSingle):
     """
-    Shift the the store by some timedelta.
+    Shift a temporal raster by some timedelta.
 
-    :param store: The source whose time is to be modified.
-    :param time: The time to shift the store, in milliseconds.
+    A positive timedelta shifts into the future and a negative timedelta shifts
+    into the past.
 
-    :type store: Store
-    :type time: integer
+    Args:
+      store (RasterBlock): The store whose timestamps are to be shifted
+      time (integer): The timedelta to shift the store, in milliseconds.
 
-    Modifies the source's properties and queries it such that its data
-    appears to be shifted towards the future in case of a positive time
-    parameter.
+    Returns:
+      RasterBlock with its timestamps shifted.
     """
 
     def __init__(self, store, time):
@@ -251,7 +252,7 @@ def _ts_to_dt(timestamp, timezone):
         timestamp = timestamp.tz_localize(timezone)
     except TypeError:
         pass
-    return timestamp.tz_convert("UTC").tz_localize(None).to_pydatetime()
+    return timestamp.tz_convert("UTC").tz_localize(None).to_pydatetime(warn=False)
 
 
 def _get_bin_label(dt, frequency, closed, label, timezone):
@@ -298,7 +299,7 @@ def _get_closest_label(dt, frequency, closed, label, timezone, side="both"):
         differences = differences[differences >= pd.Timedelta(0)]
     elif side == "left":
         differences = differences[differences <= pd.Timedelta(0)]
-    result = differences.abs().argmin()
+    result = differences.abs().idxmin()
     return _ts_to_dt(result, timezone)
 
 
@@ -333,27 +334,31 @@ def count_not_nan(x, *args, **kwargs):
 
 class TemporalAggregate(BaseSingle):
     """
-    Geoblock that resamples rasters in time.
+    Resample a raster in time.
+    
+    This operation performs temporal aggregation of rasters, for example a
+    hourly average of data that has a 5 minute resolution.. The timedelta of
+    the resulting raster is determined by the 'frequency' parameter.
 
-    :param source: The source whose time is to be modified.
-    :param frequency: the frequency to resample to, as pandas offset string
-       if this value is None, this block will return the temporal statistic
-       over the complete time range, with output timestamp at the end of the
-       source's period.
-    :param statistic: the type of statistic to perform. Can be
-      ``'sum', 'count', 'min', 'max', 'mean', 'median', 'p<percentile>'``
-    :param closed: {``None``, ``'left'``, ``'right'``}. Determines what side
-      of the interval is closed when resampling.
-    :param label: {``None``, ``'left'``, ``'right'``}. Determines what side
-      of the interval is used as output datetime.
-    :param timezone: timezone to perform the resampling in
+    Args:
+      source (RasterBlock): The input raster whose timesteps are aggregated
+      frequency (string or None): The frequency to resample to, as pandas
+        offset string (see the references below). If this value is None, this
+        block will return the temporal statistic over the complete time range,
+        with output timestamp at the end of the source raster period.
+        Defaults to None.
+      statistic (string): The type of statistic to perform. Can be one of
+        ``{"sum", "count", "min", "max", "mean", "median", "p<percentile>"}``.
+        Defaults to ``"sum"``.
+      closed (string or None): Determines what side of the interval is closed.
+        Can be ``"left"`` or ``"right"``. The default depends on the frequency.
+      label (string or None): Determines what side of the interval is closed.
+        Can be ``"left"`` or ``"right"``. The default depends on the frequency.
+      timezone (string): Timezone to perform the resampling in, defaults to
+        ``"UTC"``.
 
-    :type source: RasterBlock
-    :type frequency: string or NoneType
-    :type statistic: string
-    :type closed: string or NoneType
-    :type label: string or NoneType
-    :type timezone: string
+    Returns:
+      RasterBlock with temporally aggregated data.
 
     See also:
       https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.Series.resample.html
@@ -578,7 +583,9 @@ class TemporalAggregate(BaseSingle):
         times = time_data["time"]
 
         # convert times to a pandas series
-        series = pd.Series(index=times).tz_localize("UTC").tz_convert(timezone)
+        series = (
+            pd.Series(index=times, dtype=float).tz_localize("UTC").tz_convert(timezone)
+        )
 
         # localize the labels so we can use it as an index
         labels = labels.tz_localize("UTC").tz_convert(timezone)
@@ -632,7 +639,10 @@ class TemporalAggregate(BaseSingle):
             inds = indices[timestamp]
             if len(inds) == 0:
                 continue
-            aggregated = agg_func(values[inds], axis=0)
+            with warnings.catch_warnings():
+                # the agg_func could give use 'All-NaN slice encountered'
+                warnings.simplefilter("ignore", category=RuntimeWarning)
+                aggregated = agg_func(values[inds], axis=0)
             # keep track of NaN or inf values before casting to target dtype
             no_data_mask = ~np.isfinite(aggregated)
             # cast to target dtype
@@ -651,19 +661,26 @@ def accumulate_count_not_nan(x, *args, **kwargs):
 
 class Cumulative(BaseSingle):
     """
-    Geoblock that computes a cumulative of a raster over time.
+    Compute the cumulative of a raster over time.
 
-    :param source: The source whose time is to be modified.
-    :param statistic: the type of statistic to perform. Can be
-      ``'sum', 'count'``
-    :param frequency: the frequency at which to restart the cumulative. if this
-       value is None, the cumulative will continue indefinitely
-    :param timezone: timezone to restart the cumulative
+    Contrary to ``dask_geomodeling.raster.temporal.TemporalAggregate``, in this
+    operation the timedelta of the resulting raster equals the timedelta of the
+    input raster. Cell values are accumulated over the supplied period. At the
+    end of each period the accumulation is reset.
 
-    :type source: RasterBlock
-    :type statistic: string
-    :type frequency: string or NoneType
-    :type timezone: string
+    Args:
+      source (RasterBlock): The input raster whose timesteps are accumulated.
+      statistic (string): The type of accumulation to perform. Can be ``"sum"``
+        or ``"count"``. Defaults to ``"sum"``.
+      frequency (string or None): The period over which accumulation is
+        performed. Supply a pandas offset string (see the references below). If
+        this value is None, the accumulation will continue indefinitely.
+        Defaults to None.
+      timezone (string): Timezone in which the accumulation is performed,
+        defaults to ``"UTC"``.
+
+    Returns:
+      RasterBlock with temporally accumulated data.
 
     See also:
       https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#dateoffset-objects
@@ -797,7 +814,9 @@ class Cumulative(BaseSingle):
         closed = process_kwargs["closed"]
         label = process_kwargs["label"]
         times = (
-            pd.Series(index=time_data["time"]).tz_localize("UTC").tz_convert(timezone)
+            pd.Series(index=time_data["time"], dtype=float)
+            .tz_localize("UTC")
+            .tz_convert(timezone)
         )
 
         if frequency is None: