+++ dask_geomodeling/tests/test_geometry.py
@@ -5,7 +5,6 @@ from datetime import timedelta as Timedelta
 
 from numpy.testing import assert_almost_equal
 from osgeo import ogr
-from pandas.util.testing import assert_series_equal
 from shapely.geometry import box, Point, Polygon
 import geopandas as gpd
 import numpy as np
@@ -29,6 +28,11 @@ from dask_geomodeling.geometry import parallelize
 from dask_geomodeling.geometry import merge
 from dask_geomodeling.geometry import text
 from dask_geomodeling import geometry
+
+try:
+    from pandas.testing import assert_series_equal
+except ImportError:
+    from pandas.util.testing import assert_series_equal
 
 
 def create_geojson(abspath, polygons=10, bbox=None, ndim=2, projection="EPSG:4326"):
@@ -385,6 +389,117 @@ class TestSetOperations(unittest.TestCase):
         view = geom_operations.Area(self.empty, projection="EPSG:3857")
         result = view.get_data(**self.request)
         self.assertEqual(0, len(result))
+
+
+class TestGeometryWktSource(unittest.TestCase):
+    def setUp(self):
+        self.projection = "EPSG:28992"
+        self.geometry = shapely_transform(
+            box(135000.5, 455998, 135001.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        self.request = dict(
+            mode='intersects',
+            geometry=box(135000.5, 455998, 135001.5, 455999.5),
+            projection=self.projection,
+        )
+
+    def test_geometry_wkt_source_vals_wrong_mode(self):
+        self.request['mode'] = 'jose'
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        with self.assertRaises(ValueError) as ctx:
+            view.get_data(**self.request)
+        self.assertEqual("Unknown mode 'jose'", str(ctx.exception))
+
+    def test_geometry_wkt_source_vals(self):
+        self.request['mode'] = 'intersects'
+        assert self.geometry.intersects(self.request["geometry"])
+        assert self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"]['geometry'][0].wkt == self.geometry.wkt
+
+    def test_geometry_wkt_source_vals_intersects_not_centroid(self):
+        self.request['mode'] = 'intersects'
+        self.geometry = shapely_transform(
+            box(135001, 455998, 135002.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        assert self.geometry.intersects(self.request["geometry"])
+        assert not self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"]['geometry'][0].wkt == self.geometry.wkt
+
+    def test_geometry_wkt_source_vals_empty(self):
+        self.request['mode'] = 'intersects'
+        self.geometry = shapely_transform(
+            box(135100.5, 455998, 135101.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        assert not self.geometry.intersects(self.request["geometry"])
+        assert not self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"].empty
+
+    def test_geometry_wkt_source_vals_mode_centroid(self):
+        self.request['mode'] = 'centroid'
+        assert self.geometry.intersects(self.request["geometry"])
+        assert self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"]['geometry'][0].wkt == self.geometry.wkt
+
+    def test_geometry_wkt_source_vals_intersects_centroid_empty(self):
+        self.request['mode'] = 'centroid'
+        self.geometry = shapely_transform(
+            box(135001, 455998, 135002.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        assert self.geometry.intersects(self.request["geometry"])
+        assert not self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"].empty
+
+    def test_geometry_wkt_source_vals_centroid_empty(self):
+        self.request['mode'] = 'centroid'
+        self.geometry = shapely_transform(
+            box(135100.5, 455998, 135101.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        assert not self.geometry.intersects(self.request["geometry"])
+        assert not self.geometry.centroid.intersects(self.request["geometry"])
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"].empty
+
+    def test_geometry_wkt_source_vals_mode_extent(self):
+        self.request['mode'] = 'extent'
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual == {
+            'extent': (135000.5, 455998.0, 135001.5, 455999.5),
+            'projection': 'EPSG:28992'
+        }
+
+    def test_geometry_wkt_source_vals_extent_empty(self):
+        self.request['mode'] = 'extent'
+        self.geometry = shapely_transform(
+            box(135100.5, 455998, 135101.5, 455999.5),
+            "EPSG:28992", self.projection
+        )
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual == {'projection': 'EPSG:28992', 'extent': None}
+
+    def test_geometry_wkt_source_vals_min_size(self):
+        self.request['mode'] = 'intersects'
+        self.request['min_size'] = 2.0
+        view = geometry.GeometryWKTSource(self.geometry.wkt, self.projection)
+        actual = view.get_data(**self.request)
+        assert actual["features"].empty
 
 
 class TestConstructive(unittest.TestCase):