+++ CHANGES.rst
@@ -2,10 +2,25 @@ Changelog of dask-geomodeling
 ===================================================
 
 
-2.2.1 (unreleased)
+2.2.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added GeometryWKTSource.
+
+- Reworked the docstrings of all rasterblocks.
+
+- Renamed the 'location' parameter of raster.misc.Step to 'value'.
+
+
+2.2.1 (2020-02-04)
+------------------
+
+- Suppressed "invalid value encountered in greater than" warning in
+  ClassifyFromColumns.
+
+- Compatibility fixes for pandas 1.0.0.
+
+- Implemented raster.RasterizeWKT
 
 
 2.2.0 (2019-12-20)